RAC changes
web100-pcap, logging.h, logging.c testoptions.c, and web100srv.c all have new
code to improve the logging of snaplog and tcpdump files.  The admin can now
specify the dir (absolute or relative to BASEDIR for the log files.

In addition to the above, log files now have a new filename structure based on the
ISO time syntax.  Finally snaplog and ndttrace log files are in a subdir structure where
the YYYY/MM/DD structure is used.  All of these changes are to support archiving in the 
M-Lab infrastructure.

web100-pcap: now has a direction field, currently it just prepends c2s or s2c to the 
ndttrace file name.  

Patched server and client code to support setting of the TCP buffer.  A bug in the 
existing code prevented this option from actually working.

Bumped server version to 3.5.2