bug fix to ndttrace logging filename.  The file name is generated in the child process and then it needs to be passed back to the parent to get listed in the metadata file.
There already was a pipe and the child sent a message to the parent once the filter was
in place.  Now that message is either the ndttrace name, if logging is requested or
a "Ready" message.

also fixed bug in setting the hostname.  the meta.family varialbe is now set right after
the middlebox test is requested.

rac 7/9/09