Bug fixes-
1) zero out buffer used to receive parent-to-child "go" message.  The buffer had 
extraenous characters which caused the last test to be requeted to fail.  The
value wasn't a test number, but a string with the extraenous data attached.

2) fixed multi-client test mode, server initialization code use to be handled
once the test started, it was moved to the init stage, but the test for multi-client
mode occurred before the test_suite was initialized. Moved test for multi-client
to after initialization step.

v3.5.5 should be ready for release

RAC 7/14/09