Clean up memory leaks reported by valgrind program http://valgrind.org
Added in new error detection routine in main for() loop.  If running in multi-client
mode and the number of waiting clients (in the queue) is less than the number of mclients
then we probably missed a signal.  Test for this condition and if true, call the signal 
handler routine child_sig() to clean up.

RAC 10/13/09